Index buffer inference
Re-implemented attachment class as record
Improved some doc
Index buffer now infers data type

+++ src/main/java/org/sarge/jove/util/BufferHelper.java
@@ -36,7 +36,7 @@ public final class BufferHelper {
 	 * @param count Draw count
 	 * @return Whether index is short
 	 */
-	public static boolean isShortIndex(int count) {
+	public static boolean isShortIndex(long count) {
 		return count < BufferHelper.SHORT;
 	}
